do setup project
