+++ lib/src/modules/sample_feature/sample_item_details_view.dart
