+++ lib/src/modules/sample_feature/sample_item_list_view.dart
@@ -1,6 +1,6 @@
+import 'package:e_commerce/src/config/themes/settings_view.dart';
 import 'package:flutter/material.dart';
 
-import '../settings/settings_view.dart';
 import 'sample_item.dart';
 import 'sample_item_details_view.dart';
