+++ pubspec.yaml
@@ -30,3 +30,6 @@ flutter:
   assets:
     # Add assets from the images directory to the application.
     - assets/images/
+    - assets/logo/
+    - assets/videos/
+    - assets/fonts/
